requirements.txt
# Core dependencies
earthengine-api>=0.1.380
pandas>=2.0.0
numpy>=1.24.0

# Web application
flask>=3.0.0
flask-cors>=4.0.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data processing
pyarrow>=14.0.0  # For parquet file support

# Optional: for better performance
numba>=0.58.0
